backend/requirements.txt
fastapi
uvicorn[standard]
transformers
torch
pydantic
python-multipart
pdfplumber
fuzzywuzzy
passlib
langchain
faiss-cpu
langchain
sentence-transformers
pymongo
bcrypt==4.3.0
jose
python-dotenv
langchain-google-genai
tdqm
langchain-openai
langchain-core
sqlparse
pandas
attrs
typing